create Extention file to transfer the lable , stackview, NavigationBar, and extra..
