+++ TodoListPro/ViewModel/ViewModel.swift
@@ -46,21 +46,6 @@ class TodoViewModel {
         }
     }
     
-
-    
-    func navigationConfig(nav: UIViewController) {
-        nav.title = "List"
-        let nadv = nav.navigationController?.navigationBar
-        nadv?.backgroundColor = .gray
-    }
-  
-    func addItem(viewController: UIViewController){
-        let vc = AddInToTodoListViewController()
-        let navVC = UINavigationController(rootViewController: vc)
-        navVC.modalPresentationStyle = .fullScreen
-        viewController.present(navVC, animated: true)
-    }
-    
      func refresh(tableView: UITableView) {
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
             tableView.refreshControl?.endRefreshing()
@@ -84,26 +69,4 @@ class TodoViewModel {
        deleteFormList(the: theDeletOpetion)
        todoArray.remove(at: indexPath.row)
     }
-
-    func activityIndicator(activity: UIActivityIndicatorView, view: UIView) {
-        activity.center = view.center
-        activity.hidesWhenStopped = true
-        activity.style = .large
-        activity.color = UIColor.black
-        view.addSubview(activity)
-    }
-    
-    func startActivity(activity: UIActivityIndicatorView, view: UIView) {
-        activity.startAnimating()
-        view.isUserInteractionEnabled = false
-    }
-    func stopActivity(activity: UIActivityIndicatorView, view: UIView) {
-        activity.stopAnimating()
-        view.isUserInteractionEnabled = true
-    }
-
-    
-  
-    
-    
 }
